src/gmail_impl/pyproject.toml
[project]
name = "gmail-impl"
version = "0.1.0"
description = "Gmail implementation of the EmailClient protocol with dependency injection"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "email-api>=0.1.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.1.0",
    "google-api-python-client>=2.100.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gmail_impl"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "--strict-markers",
    "--strict-config", 
    "--cov=src/gmail_impl",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=85",
]
testpaths = ["tests"]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
]
asyncio_mode = "auto"

# Coverage configuration  
[tool.coverage.run]
source = ["src/gmail_impl"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "\\.\\.\\.",
] 